Started testing the views of `apps.sitenews`
I've tested that news, news lists and year archives return an HTTP
200 code when hitting their respective URLs and the correct templates
are rendered.

In the case of non-existing news, I've tested that an HTTP 404 error is
raised. I also created the `richard/templates/404.html` empty template.

Finally, I added a couple of lines to `.gitignore`: one to ignore
sqlite3 .db databases (which I use locally) and a `.env` file that I use
for environment-specific settings.