Break up documentation
Previously the documentation was in two parts, but if you looked at
the documentation in a web browser, the contributor guide was outside
of the viewport.

This moves things around so that the front page of the docs does a
better job of explaining that the guide is in two parts, links to the
parts, and explains what each part is.

Also, it puts the files into guide-specific directories.