Change USE_TZ to False
This shuts off the timezone-aware stuff. It's not clear that timezone-aware
dates and datetimes really helps a video index site where you can't really
be sure what timezone the original data is in. Given that and the fact
that it makes everything really spammy with all the warnings, it seems
prudent to just shut it off.