77. Admins have write access to the API
Authentication is done with username and api key. If that fails, the
user only gains read access to the resources.

Addings videos works, but we may want to consider to simplify the
interface. Right now, relations like category and speaker need to be
added by putting the URI of the resource (e.g. '/api/v1/category/1/')
into the payload.
I'm thinking of passing in just names and let richard figure it out.