Fix sitenews feed; tidy up
A sitenews item consists of an "above-the-fold" and "below-the-fold" sections
which together make up the sitenews item. So it makes sense to have both
in the feed.

This also tidies up some of the code turning things into unicodes that
should be unicodes and fixing the __repr__ so they're less susceptible
to unicode errors.