70. Add test for opensearch suggestions
For this test we need a search index that we can query, therefore in
richard/settings_test.py we are pointing whoosh to a test directory that
we can remove after the test.
Suggestions are activated when running tests, this helps when testing
the suggestions view because its URL is included in the urlconf.