89. Link to specific timecode of videos wrapped in unisubs
Similar to YouTube, people can link to a specific part of a video. We
can use this (at pyvideo) to provide direct links for example to the
different presentations in lightning talks.

For example this URL will start playing at 15 minutes and 30s.

http://pyvideo.org/video/956/keynote-guido-van-rossum#t=15m30s

Right now, this is only working for videos that are managed by
UniversalSubtitles, because they expose a very helpful `playAt` function
that hopefully works across all their supported players.
I only tested this with YouTube videos so far.