106. Add a form to submit video suggestions
The form will be shown as dialog with JavaScript. Form rendering and
validation is still done by Django on the server side, JavaScript
fetches the HTML from the server when the dialog is opened and for
each submit.
Later we could probably improve here by doing additional client-side
validation to save some requests.