Tweak how related urls are shown
I added a .display() method to RelatedUrl that dictates how it gets
shown. I couldn't think of a better method name. It'd be nice to
conform to some kind of standard method name for this sort of thing.

I also moved the related urls section up in the page so it comes right
after the summary. Some pyvideo videos have really long descriptions and
I think it's probably a common use case for someone to go to the video
page looking for a related url (slides, related blog post, project
url, ...) and scrolling all the way down through the description seems
dull. If this turns out to be a bad use case assumption, we can change
it back.