Issue 130. Rework suggestions interface
This reworks and centralizes suggestions so that clicking on "Submit"
in the navbar brings you to the suggestions page where you can see
the list of open suggestions and submit a new suggestion.

This also tweaks the language again in an attempt to make it less
pyvideo.org-centric and more general.