Change GET video to return category title
We want to use the category title so that the video json data has all
the information in it and doesn't refer to other api things to get
additional information.

This does create the minor weirdness that given a category title, you
can't get the category data from the api easily. We can fix that later
probably with a filter on categories.